src/test/java/epam/library/controller/LibraryServletTest.java
package epam.library.controller;

import epam.library.controller.LibraryServlet;
import org.junit.Test;

import javax.servlet.RequestDispatcher;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

import static org.mockito.Mockito.*;

public class LibraryServletTest {
    private final static String path = "/WEB-INF/view/libraryPage.jsp";

    @Test
    public void whenCallDoGetThenServletReturnLibraryPage() throws ServletException, IOException {

        final LibraryServlet servlet = new  LibraryServlet();

        final HttpServletRequest request = mock(HttpServletRequest.class);
        final HttpServletResponse response = mock(HttpServletResponse.class);
        final RequestDispatcher dispatcher = mock(RequestDispatcher.class);

        when(request.getRequestDispatcher(path)).thenReturn(dispatcher);

        servlet.doGet(request, response);

        verify(request, times(1)).getRequestDispatcher(path);
        verify(request, never()).getSession();
        verify(dispatcher).forward(request, response);
    }

}
